Split simple_dds module.
