+++ simulator/simple_dds/entity.py
@@ -1,0 +1,15 @@
+class Entity:
+
+    def __init__(self):
+        self.instance_handle = 0
+    
+    def set_instance_handle(self, handle):
+        if self.instance_handle != 0:
+            raise RuntimeError("DDS Instance already has a handle.")
+        else:
+            self.instance_handle = handle
+
+    def get_instance_handle(self):
+        if self.instance_handle == 0:
+            raise RuntimeError("DDS Instance has not been assigned a handle")
+        return self.instance_handle
